generic write on FetchProgress
