+++ utils/net_utils/Cargo.toml
@@ -10,3 +10,6 @@ url = "2.1.1"
 log = "0.4.8"
 thiserror = "1.0"
 pbr = "1.0.3"
+pin-project-lite = "0.1"
+async-std = { version = "1.6.3", features = ["attributes"] }
+futures = "0.3.5"
