+++ utils/paramfetch/Cargo.toml
@@ -17,3 +17,4 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 isahc = "0.9.11"
 url = "2.1.1"
+net_utils = { path = "../net_utils" }
